multicall: split option handling routines out to multicall-options.c
