+++ Makefile
@@ -39,7 +39,9 @@ LIBIFUPDOWN_SRC = \
 LIBIFUPDOWN_OBJ = ${LIBIFUPDOWN_SRC:.c=.o}
 LIBIFUPDOWN_LIB = libifupdown.a
 
-MULTICALL_SRC = cmd/multicall.c
+MULTICALL_SRC = \
+	cmd/multicall.c \
+	cmd/multicall-options.c
 MULTICALL_OBJ = ${MULTICALL_SRC:.c=.o}
 MULTICALL = ifupdown
